Move name resolution into the search module
